finish message screen at all

+++ app/src/main/java/com/octopus/socialnetwork/ui/screen/message_screen/uistate/RecentMessagesUiState.kt
@@ -3,8 +3,9 @@ package com.octopus.socialnetwork.ui.screen.message_screen.uistate
 data class RecentMessagesUiState(
     val idOfSender: Int = 0,
     val nameOfSender: String = "",
-    val timeOfLastSend: Int = 0,
+    val timeOfLastSend: Long = 0,
     val avatar: String = "",
     val viewed: String = "",
     val message: String = "",
+
 )
